web/backend/requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.1
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.30
eventlet==0.36.1
python-dotenv==1.0.1
bcrypt==4.1.3
reportlab==4.2.0
